Update 02x_Migration_Guide_Self_Upgraders.txt
Looks good.  I think the portion directing the user to the right section will be a list, with links taking them to the correct section of the page.  That or we can make a handful of pages to simplify the view for the user, but this is a decision for after added to docs.